Don't execute :p and :html commands when loading libraries.
This tweak allows libraries to be literate programs with embedded
examples, because the example will only run when rendering the library
with `dex script`, not when loading it for use.

Also uncomment embedded examples in the `diagram` library, and add
libraries to quine tests to make sure all their embedded examples
execute without error.

Co-authored-by: Dougal Maclaurin <dougalm@google.com>