============== WIP ==============
Handle Ix dict hoisting in the same way we handle function specialization.

The problems of Ix dict hoisting and function specialization are very similar.
In both cases, we want to emit named chunks of code to avoid the code blowups
we'd get if we inlined them instead. We want to de-dup where possible by using
the same names in several places. We need a way to run the rest of the compiler
passes on the chunks. And we don't learn which chunks we need until
simplification, where we finally have concrete type and dictionary arguments.

Previously, we handled Ix dict hoisting and function specialization in different
ways. For function specialization, we'd gather the required specializations
during simplification. On returning to the top level, we'd lower the specialized
functions all the way to object code, before proceeding with the rest of the
passes on the original code. For Ix dict hoisting, we'd gather the required Ix
dicts in a separate pass, and then keep these dicts alongside the original code
through the subsequent passes, until Imp where they get inlined and translated.
This meant that each pass needed to handle Ix dicts specially. It also meant
that we didn't handle Ix dicts at all for standalone functions.

This patch treats Ix dicts just like function specialization, except that
instead of going all the way to object code, we only go as far as Imp, at which
point we inline them (so that LLVM can do further optimizations). This makes the
code simpler (less XXX loc) and it makes it possible to merge the separate block
evaluation and function compilation top-level paths, at which point we'll be
able to do lowering and optimization for standalone functions.