Propagate more name hints.
This change tries to increase the distance that user-supplied names
make it through compiler passes.  It's far from complete, but suffices
to get simplification to preserve the name given in the checked-in
test case, which it didn't before.

The place where user-supplied names come from is decls (that were
present in the source program).  Such a name names the value returned
from the RHS of that decl.  So, this PR collects the name hint from a
decl undergoing processing, passes it to the function processing the
RHS, and tweaks said function to, if it emits a decl representing the
value of that RHS, propagate said name hint to that emission.

Why do this now?  Because I want to write unit tests for the upcoming
occurrence analysis that query the occurrence-analyzed IR for the
annotation that goes with a variable of interest, and I want to name
that variable using the surface name that appears in the string in the
unit test.