Add a role system and use it to guide generalization.
The idea is that each parameter to a type constructor, class, or instance, must
have one of three roles:
  1. type parameter role        (parameters of type `Type`)
  2. data parameter role        (parameters that are runtime-representable)
  3. dictionary parameter role  (parameters of type `DictTy _`)

When we generalize the arguments for a function specialization, we handle
parameters according to their role. Data parameters become additional runtime
arguments. Type parameters are traversed recursively. And dictionary parameters
are generalized according to the required more general type.

For example, imagine we want to generalize the application of `sum : {a} [Ix a]
(a=>Float) -> Float` applied to `(Fin 10)` and `IxFinInstance 10`. We traverse
the type parameter, `Fin 10`, and find that it has a data parameter, `n:Nat`,
which is instantiated at `10` here. We generalize it to `n`, so we have `\n:Nat.
sum (Fin n)`. Then we consider the dictionary, `IxFinInstance 10`. With `sum`
specialized at `Fin n` instead of `Fin 10`, we now require that the dictionary
has type `Ix (Fin n)` instead of `Ix (Fin (10))`. We assume (and require) that
instances are parametric in data parameters, so we know that we can use the same
synthesis tree and we just need to update the parameters. We swap out data
parameters in the instance for inference variables, `IxFinInstance ?1`, and
unify the type, `Ix (Fin ?1)`, with the required `Ix (Fin n)`, which is solved
by `?1 = n`. We end up generating a generalized-specialized top-level function
defined by `\n. sum (Fin n) (IxFinInstance n)` as we hoped.

This change generalizes the function `generalizeType` which previously just
special-cased `Fin n` and `Ix (Fin n)` and didn't handle other types and classes
at all. It also lets us hoist more general ix dictionaries (e.g. `\n. Ix (Fin
n)` instead of `Ix (Fin 10)`, `Ix (Fin 13)`, etc.), which should produce more
cache hits, and makes it more reasonable to emit them to the top level.

With this change, we're formalizing two restrictions which were previously just
unimplemented/crashy holes in the compiler:
  1. instances must be parametric over all data arguments. You can't define an
     instance like `Ix (Fin 10)`. It has to be `{n:Nat} (Ix (Fin n))` instead.
  2. type constructor parameters must be one of type/data/dict, so we can't
     have, e.g., functions as type constructor parameters.

This commit doesn't actually enforce these invariants - either the
parametricity or the data/dict/type constraint on type parameters. We should do
that as a follow-up.