Fix a bug I accidentally introduced in 6772b86657.
The problem was that these `SpecializedDictBinding`s emissions can contain names
requiring specialization, and that wasn't captured by our free var querying. So
I made the change that tracks which specializations are needed explicitly. (We'd
been meaning to do this anyway.) As a result, I had to generalize
`HoistingTopBuilder` to allow emissions other than `TopEnvFrag`.