Move Ix hoisting to simplification pass
We often use the `substM` shortcut instead of `simplifyAtom` for things like
types and dictionaries. We had to expand these into full traversals to make sure
we find all the dictionaries. It sure feels like a lot of boilerplate. It'll be
useful for the upcoming newtype stripping, so I think it's still worth it, but
I wish we had better traversal helpers.