Cache specialized/hoisted dicts properly.
Previously we had been caching the methods but not the specialized dicts
themselves. We would emit a new one each time we encountered an `Ix` dict.
Unsurprisingly, we burned through names at a crazy rate, to the point where we
ran out of string names and spilled over to the numerical ones, which messed up
our quine tests.

I took the opportunity to simplify the dict cache by storing the simplified
methods alongside the `SpecializedDictName` binding, in a
`Maybe [NaryLamExpr n]` field. It's initially `Nothing`, when the dicionaries
are first hoisted during simplification. Then it gets populated with the
simplified methods after we return to the top level and process the "required
specializations". I think this pattern is much easier to follow and we should
use it for specialized top functions too.