Add an `IR` type-level parameter to Atom/Expr/Block.
For now this doesn't do anything, but the idea is to use it to distinguish the
different IR variants and statically encode which constructors they can each
use. This commit just adds the parameter to all the types and tries to change as
little as possible otherwise.

I used `unsafeCoerceIRE` (of type `e r n -> e r' n`) liberally to change the IR
parameter as needed. You might expect this wouldn't be necessary because we only
have a single inhabitant of `IR` and we don't have any constraints on the
constructors yet. But there are some places in the compiler where we like to be
polymorphic in `r` (e.g. type checking, builder, generic traversal) and there
are other places where we have to be monomorphic (most concrete passes). So we
often need to use `unsafeCoerceIRE` to turn e.g. `Atom r n` into `Atom CoreIR n`.

It looks like the hardest thing will be handling the terms associated with
top-level names. These will be monomorphic in the IR parameter, which means
we'll only be able to use them in a particular IR. For example, the definitions
of instance methods may be expressed in the core IR, so they won't be able to be
used in the simplified IR without explicit lowering. This is a good thing and
it's exactly why we're doing this. But it will require some changes to the
core data types.