Use destination-passing style at the top level, not just for internal calls.
This avoids the atom reconstruction we had to do after the imp pass, which was
always a pain to maintain. It should let us merge the different compilation
paths into one, standardizing on FOMO functions as the core compilation unit.

One big difference is that we now allocate the destinations within the Haskell
runtime rather than within generated code. This means we rely more on the
interpreter than we used to, including for generating the required allocation
sizes. I think this is ok because (1) the interpreter (acting on simplified IR)
should be complete and (2) codegen for code that's executed exactly once,
without any loops, is wasteful anyway.