Use an entirely type-determined data layout in Imp
Previously, we used table lambdas to represent table values and table dests.
These table lambdas usually followed a predictable pattern but since they could
in principle contain arbitrary code we had to defensively copy them before
passing them to a function. With this change, each post-simplification type
corresponds to a fixed number of scalar values (including pointers) whose
meaning can be understand if you have access to the type.

This also changes the calling convention for entry functions and interanl
functions. Entry functions now return the scalar values corresponding to their
canonical results and there's no longer a value-dependent reconstruction.
Internal functions now pass arguments by value instead of by reference as they
did previously. Results are still returned in destination passing style.

Currently, we still reconstruct a table lambda when we return to the top level
which means that subsequent top-level blocks have to generate loops to bring it
back to canonical form again. This is silly, but I wanted to minimize changes to
existing code (it's already a big patch) so I didn't want canonical atoms to
leak beyond the Imp pass. As a follow-up we can maintain the data in canonical
form at the top level, which is both simpler and more efficient.

An important side benefit of this change is that we can now use `memcpy` instead
of looping through indices when we need to copy an array. This should yield both
compile-time and run-time speedups.