Traverse the type of a dict hole to be synthesized when synthesizing.
Failing to do so created a corner case where an unsynthesized Ix dict
could be left in the IR after synthesis.  This was masked because that
dict was never accessed down-stream, but a version of the inliner
moved it to a place where it _was_ accessed, revealing the existence
of a bug.