Statically disallow lambda/pi in SimpIR.
This required a (well overdue) update to the way we handle the lambda-like
parameters of our primitive higher-order functions. Previously, our HOFs took
arbitrary Atoms as arguments but we vaguely expected that they would be
syntactic `Lam` atoms. Now we use a `LamExpr n`, which is just a value of an
(n-ary) lambda without the arrows, effects and so on, and we can forbid the
`Lam` and `Pi` constructors in the simplified IR even though we still have HOFs.