Generalize `StringSetIx` to any element type, and rename it `Element`.
Usage example:

people = to_set <a bunch of strings>
Person = Element String people
for i:Person.
  ...

At runtime, `Person` is an integer indexing into the array backing the
set.  You can compare them for equality and ordering, and use them as
indices into other arrays (TODO and hash them) without having to mess
with the actual strings; but the strings are there if you ever need
them.

Tagline: The code book is in the type.