Fix compile times reported by %bench.
Compilation time is computed as "total time - time spent evaluating".
There were two distinct evaluations that %bench did whose time it did
not include in the total: the unconditional evaluation that happens
whether benchmarking or not, and the test evaluation that benchmarking
itself does to estimate how many evaluations it can complete in two
seconds.

Since it was convenient, I also include the latter in the average for
reporting average run time.