Remove DictCon, DictExpr, and ProjMethod from SimpIR.
Ix dicts are still used in Imp, but we ensure that they only occur in their
specialized, simplified form.

Sadly, some of our unsafeCoerceIREs are starting to cause runtime problems. I'm
going to follow this up with the newtype stripping which should let us remove
many of them and we'll see if that fixes things.