Strip newtypes during simplification
With this patch, SimpIR no longer contains: newtypes (including Fin, Nat,
records, variants, user-defined ADTs), functions, dictionaries, table
lambdas, or method projections.

The approach was to add a `SimpInCore` constructor to `Atom` for embedding
simplified names in the transitional SimpToCoreIR atoms, optionally coerced to
some richer type. It mostly works, but there are still some bugs, including
a fundamental one to do with non-data reconstructions. Checking this in for now
as a checkpoint.