Remove interpreter.
Since we moved to codegen-based printing the only remaining use of the
interpreter is during inference. There, we use use it for (1) top-level
pattern-matching and (2) to figure out the size of non-Fin table literals. We
can handle (1) by codegen and a telescope. And (2) is a lot of complexity for a
small feature, especially one that was always going to be incomplete anyway.
Instead we can just use `coerce_table` which does a runtime check on the size.