Add an IR parameter to effects and effect rows.
This is in preparation for tagging names with IR parameters. I also made
`UEffectRow` and `EffectRow` distinct ADTs rather than just different
instantiations of `EffectRowP`. The parameterization was getting too awkward. I
also removed the `Maybe` wrapper on the heap parameter because we don't need it
now that we have a `HeapVal` value to serve the role of `Nothing`.