Add an `IRRep` constraint to obtain a runtime representation of IR parameters.
`IRRep` is like `Color`, but for `IR` isntead of `C`. I tried getting rid of `Color`
altogether but there are a few places we need it. One place is serialization.
Another is when we query free vars of a particular color using `freeVarsList`. I
think we can probably work around both of these but it would be an (even) bigger
change.