Bundle the `Transpose` primitive with its cotangent argument.
Previously we had the curried version of the primitive, where the result was the
transposed function itself. But SimpIR doesn't have functions. That used to be
fine because we'd beta-reduce the result soon enough anway. But it's a problem
now that we're being strict about the CoreIR/SimpIR distinction. Supplying the
cotangent argument fixes it, and currying can be done in user-space if desired.