Obey CoreIR/SimpIR distinction in TabLam/ACase reconstruction terms.
TabLam needs to have a SimpIR binder and a CoreIR body, while ACase
needs a SimpIR scrutinee, SimpIR binders, and a CoreIR body. I put both of these
under a three-case `SimpInCore` type and removed them from `Atom`.

Since I removed `TabLam`, the trivial loop unrolling doesn't work anymore. If
it's an important optimization we should find another place for it. We could do
it during simplification (using the table lambda within `SimpInCore`) or we
could just handle it in the inliner.