Make linearize produce SimpIR instead of CoreIR.
Previously we would emit CoreIR from linearization so that we'd emit tangent
functions as actual functions and rely on our general-purpose
defunctionalization machinery to carry them to their use site. But it turns out
we were being sloppy about the SimpIR->CoreIR lifting. In particular,
linearization carried an *IR changing* substitition, mapping SimpIR names to
CoreIR names. We could have made that work but instead we decided to just
reimplement a specialized version of defunctionalization during the
linearization pass, so that linearization becomes a SimpIR->SimpIR
transformation. This might be a good thing anyway because during linearization
we have more information about where the functions will be used.

Two other small changes:
  1. Bundle the `Linearize` primitive with its (first) argument, just as we did
     with `Transpose` recently. This just lets us treat the result of linearize
     as a first-order function rather than a second-order one, which is a bit
     tidier.
  2. Add a `LiftSimpFun` case to `SimpInCore` for carrying the result of
     linearization.

Co-authored-by: Alexey Radul <axch@google.com>