Fix bug in Lower where we were doing an ill-typed substitution.
As we hoisted alloc decls in `Seq` we were extending the substition with
projections of the loop binder. But we hadn't yet updated the loop binder's type
to reflect the additional dests. We couldn't even do that if we'd wanted to
because we didn't know ahead of time what extra dests we'd need. This problem
existed before but it was only exposed with the recent change that gave the
substition traversal access to the type environment. The fix is to just hoist
the decls first and then add the extra dests to the loop carry afterwards.

Co-authored-by: Alexey Radul <axch@google.com>