Fix worker thread ending to a correct-but-inefficient scheme.
To wit, there is a global variable for whether the computation is
over.  It is set to `false` at the beginning, and the (unique) last
Work in the computation sets it to `true` before exiting.  All the
worker threads keep trying to steal until the global flag becomes
`true`, regardless of how often their thefts fail.

This could be optimized in lots of ways, but at least this much is
correct.