Implement and test synthesis for the Data constraint.
The idea is that Data is supposed to tag types that "are data", in the
sense of having a runtime representation.  We know how to check in the
compiler whether a type meets this requirement, so we implement that
as a code path in dictionary synthesis.

We don't want the user to implement Data themselves, because the
compiler assumes its invariant in various places.  Indeed, Issue #1150
references a variety of the ways in which we use it.