Add the Data constraint to Ref, and downstream.
It seems that we have also been assuming in various places that
classes like Eq, Add, Mul, and Ix also imply Data, in the sense that
types with those classes are suitable for being stored in references.
Added Data as a superclass to those, pending a design decision on the
subject in Issue #1230.

Fixes #609.