Remove "allowed effects" from the Env we carry everywhere.
We only actually use the allowd effects in two places: inference and type
checking. Tracking them everywhere else hurts performance because we have to
consider extending the allowed effects every time we emit a binding, which is a
lot. Since we added the IR parameters, the effects carried in the Env are also
ill-typed, because they're treated as CoreIR even though we can have SimpIR
effects too. We had to use unsafeCoerceIRE to work around that. After this
change, the only use of unsafeCoerceIRE is in pretty-printing.