Make the check for recursive method definitions less syntactic.
This is achieved by encoding inside a `ClassArrow` which class
methods the expression "to the right" of the arrow requires
access to. The arrow can require `Full` access, in which case
all methods are accessible in the subordinate expression ("to
the right") of the arrow. Alternatively, an arrow can require
`Partial` access, in which case the `ClassArrow` also carries
a list of method indices of those methods in the class that
the arrow requires access to.

Note that the surface language has not changed: when a user
specifies a class as a (function) argument, the `ClassArrow`
resulting from this is assumed to require `Full` access.

The only place where `Partial` arrows are constructed in the
compiler is when building the types for method projections
(in function `makeMethodGetter`). A method projection requires
access to exactly one method of a class/dictionary.

During type inference, the method accesses that are required by
a `ClassArrow` are fed through to `DictHole`s. Later, when an
instance is synthesized for the `DictHole`, it is checked that
the synthesized instance has at least as many methods as the
`DictHole` requires access to.