Split TopLevel.evalLLVM into separate compile and run phases.
The run phase (now named callEntryFun) still requires the function to
be compiled with EntryFunCC, and there is not yet a code path to so
compile a function that accepts arguments (as opposed to a block).
Ergo, the arguments it nominally accepts are somewhat moot.

We also change the boundary of `toCFunction` not to include
`emitObjFile`, so the new `packageLLVMCallable` (which is the compile
phase of the old `evalLLVM`) can reuse it.