Simplify the `RequiredMethodAccess` data type.
Previously, the `Partial` data constructor used carried a list of
method indices that were used to indicate the class/instance methods
a `ClassArrow` requires access to.

This commit changes the `Partial` data constuctor to take only a single
`Int` as its argument. This `Int` indicates how many class/instance
methods are accessible. Encoding in a `ClassArrow` only how many class/
instance methods must be accessible (instead of encoding the precise
indices of the required methods) is sufficient because
  - the `Int` argument of `Partial` is used only to control method
    access in partially defined instances, and
  - instance method definitions are (internally) processed in the order
    in which methods are defined in the corresponding class/interface.
Hence, it suffices to compare the number of already defined instance
methods with the number of required class/interface methods.