Update diagram.dx to the new syntax.
I changed the API slightly, to work well with | splicing arguments into applications.

To wit, where the API used to be
  trasform(args) -> (Diagram) -> Diagram
I changed it to
  transform(Diagram, args) -> Diagram

The former worked well as a pipeline using |>, where each element of
the pipeline had to be a unary function.  The new version works well
with |, which creates pipelines by adding the LHS as the first
argument of the RHS, if the latter is an application.