Improve "ambiguous type" errors by tracking the purpose of each inference variable.
Instead of the dreaded error:

  ambiguous type variables: [_.1]

We now get:

  Couldn't infer type of unannotated binder b

or

  Couldn't infer implicit argument b of foo

Where `b` and `foo` are actual source names, not just names invented by the
compiler. These two cases -- unannotated binders and implicit args -- account
for most of the inference variables we create, so those were the error messages
I focussed on. Now that the machinery is in place we can improve the rest of
them incrementally as they occur.