Use the same name for the type constructor and the data constructor of a struct
This "pun" is pretty conventional. It worked well for tuples and it feels less
clumsy than writing `Foo.new`. We default to treating the name as a data
constructor unless we know from the context that we're expecting a type. It
introduces one footgun, which is that you get a confusing error message if you
try to make a type alias without an annotation, e.g. `Person = Element names` vs
`Person : Type = Element names`. We should add a way to include relevant
information like "Element was interpreted as a data constructor".