Simplify vestigial destination passing implementation in Imp.hs.
To wit, the only time translateExpr could ever get a non-Nothing dest
was when translating the result of a block in translateBlock, which is
statically always an Atom.  So specialize that case to a new
translateAtom, and remove the Maybe Dest argument from translateExpr.
All of translateExpr's helpers now also statically get Nothing dests,
so remove those Maybe Dest arguments as well.