Remove type/effect annotations from `Block`.
Blocks are now just a bunch of decls followed by an atom. The type and effects
information is carried by the context -- e.g. a pi type or the other arguments
to a built-in HOF. The reason is that we're about to add decls to binders and
this way we can treat binder-decls block-decls uniformly.